pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cnm"
version = "0.1.0"
description = "Compositional Network Model for Chinese Pre-trained Language Models"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10,<3.12"
authors = [
    {name = "Eric Yan"},
]
keywords = [
    "chinese",
    "nlp",
    "bert",
    "transformers",
    "compositional",
    "ideographic",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.2.0",
    "transformers>=4.37.0",
    "datasets>=2.16.0",
    "tokenizers>=0.15.0",
    "accelerate>=0.26.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "jieba>=0.42.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "wandb>=0.16.0",
    "evaluate>=0.4.0",
    "sentencepiece>=0.1.99",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]
corpus = [
    "mwparserfromhell>=0.6.5",
    "wikitextparser>=0.55.0",
]

[project.scripts]
cnm-download-ids = "cnm.scripts.download_ids:main"
cnm-prepare-ids = "cnm.scripts.prepare_ids:main"
cnm-prepare-corpus = "cnm.scripts.prepare_corpus:main"
cnm-pretrain = "cnm.scripts.pretrain:main"
cnm-finetune = "cnm.scripts.finetune:main"
cnm-evaluate = "cnm.scripts.evaluate:main"

[project.urls]
Repository = "https://github.com/ericyan/cnm"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cnm = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["cnm"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=cnm --cov-report=term-missing"
